+++ svelte.config.js
@@ -9,7 +9,8 @@ const config = {
       precompress: true,
     }),
     csrf: {
-      checkOrigin: process.env.NODE_ENV !== "development",
+      checkOrigin: false,
+      // checkOrigin: process.env.NODE_ENV !== "development",
     },
     alias: {
       "$smtp": "./smtp.config.cjs",